requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-filter==22.1
django-jazzmin==2.6.0
django-photologue==3.15.1
django-rest-framework==0.1.0
django-sortedm2m==3.1.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
ExifRead==3.0.0
factory-boy==3.2.1
Faker==16.6.1
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
pytest==7.2.1
pytest-django==4.5.2
pytest-sugar==0.9.6
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
